Fixed tests to work with GitHub Actions
Don't run the ProgressReporter test cases and changed
the SndFileAudioFileReader and Mp3AudioFileReader tests
to avoid checking the ProgressReporter output.

This is a workaround for an issue with running the tests
under GitHub Actions.

When run under GitHub Actions, S_ISFIFO() returns 1 whereas
when run locally it returns 0.

This affects the ProgressReporter class, which uses this
check to enable or disable percentage progress messages.